Add flask_restx api to OCEAN
