+++ ocean/app/main.py
@@ -23,7 +23,7 @@ from datetime import datetime
 from jinja2 import Environment
 from flask import Flask, flash, redirect, render_template, request, url_for, session as client_session, g
 from flask_login import LoginManager, login_user,login_required, current_user, login_manager, logout_user
-from flask_restful import Resource, Api, reqparse, fields, marshal_with, abort
+from flask_restx import Api
 
 # Local application imports
 from . import models
@@ -40,7 +40,7 @@ from .routes.routes_encounter import routes_encounter
 from .routes.routes_datahub import routes_datahub
 from .routes.routes_healthcentre import routes_healthcentre
 from .routes.routes_filespace import routes_filespace 
-
+from .routes.api import blueprint as api
 
 def check_interfaces():
     """
@@ -73,15 +73,7 @@ def check_file_space():
 def create_app(config_class):
     app = Flask(__name__)
     api = Api(app)
-
-    user_args = reqparse.RequestParser()
-    user_args.add_argument('name', type=str, required=True, help="Name cannot be blank")
-    user_args.add_argument('email', type=str, required=True, help="Email cannot be blank")
-
     check_interfaces()
-
-    
-
     ROUTE_PREFIX = '/ocean'
 
     # Set up a custom login manager for the web app
@@ -156,9 +148,8 @@ def create_app(config_class):
     app.register_blueprint(routes_datahub, url_prefix=ROUTE_PREFIX)
     app.register_blueprint(routes_healthcentre, url_prefix=ROUTE_PREFIX)
     app.register_blueprint(routes_filespace, url_prefix=ROUTE_PREFIX)
-    from .routes.api import test_api
-    test_api.register(api, prefix=ROUTE_PREFIX)
-    # app.register_blueprint(test_api, url_prefix=ROUTE_PREFIX)
+    
+    app.register_blueprint(api, url_prefix=ROUTE_PREFIX + "/api/")
 
     try:
         logger.info(check_file_space())
