--- ocean/app/routes/api/test_api.py
@@ -1,16 +1,0 @@
-from flask_restful import Resource, Api, reqparse, fields, marshal_with, abort
-from flask import send_from_directory
-
-from ... import database_handler
-from . import resources
-
-def register(api, prefix):
-    prefix = str(prefix)
-    if str(prefix).endswith('/'):
-        prefix = str(prefix)[:-1]
-    if not str(prefix).endswith('api'):
-        prefix = prefix + "/" + 'api'
-    
-    api.add_resource(resources.SelectionResource, prefix + '/selections/')
-    api.add_resource(resources.FileResource, prefix + '/file/')
-    api.add_resource(resources.SpectrogramResource, prefix + '/spectrogram/')
