+++ requirements.txt
@@ -1,6 +1,10 @@
+alabaster==1.0.0
 amqp==5.2.0
+aniso8601==10.0.0
 annotated-types==0.7.0
+attrs==25.1.0
 audioread==3.0.1
+babel==2.16.0
 bidict==0.23.1
 billiard==4.2.1
 blinker==1.8.2
@@ -18,12 +22,15 @@ contourpy==1.3.0
 coverage==7.6.1
 cycler==0.12.1
 decorator==5.1.1
+docutils==0.21.2
 et-xmlfile==1.1.0
 exceptiongroup==1.2.2
 factory_boy==3.3.1
 Faker==33.1.0
 Flask==3.0.3
+Flask-Admin==1.6.1
 Flask-Login==0.6.3
+flask-restx==1.3.0
 Flask-SQLAlchemy==3.1.1
 Flask-Testing==0.8.1
 fonttools==4.53.1
@@ -32,15 +39,21 @@ h11==0.14.0
 hurry==1.1
 hurry.filesize==0.9
 idna==3.8
+imagesize==1.4.1
+importlib_resources==6.5.2
 iniconfig==2.0.0
 itsdangerous==2.2.0
 Jinja2==3.1.5
 joblib==1.4.2
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
 kiwisolver==1.4.5
 kombu==5.4.2
 lazy_loader==0.4
 librosa==0.10.2.post1
 llvmlite==0.43.0
+Mako==1.3.8
+Markdown==3.7
 MarkupSafe==2.1.5
 matplotlib==3.9.2
 msgpack==1.0.8
@@ -52,6 +65,7 @@ numpy==2.0.2
 openpyxl==3.1.5
 packaging==24.1
 pandas==2.2.2
+pdoc3==0.11.5
 pillow==10.4.0
 platformdirs==4.2.2
 pluggy==1.5.0
@@ -61,22 +75,37 @@ psutil==6.0.0
 pycparser==2.22
 pydantic==2.9.0
 pydantic_core==2.23.2
+Pygments==2.19.1
 PyMySQL==1.1.1
+pynvim==0.5.2
 pyparsing==3.1.4
 pytest==8.3.2
+pytest-cov==6.0.0
 pytest-flask==1.3.0
 python-dateutil==2.9.0.post0
 python-engineio==4.10.1
 python-socketio==5.11.4
 pytz==2024.1
+referencing==0.36.2
 requests==2.32.3
+rpds-py==0.22.3
 scikit-learn==1.5.1
 scipy==1.14.1
 setuptools==75.6.0
 simple-websocket==1.1.0
 six==1.16.0
+snowballstemmer==2.2.0
 soundfile==0.12.1
 soxr==0.5.0.post1
+Sphinx==8.1.3
+sphinx-rtd-theme==3.0.2
+sphinxcontrib-applehelp==2.0.0
+sphinxcontrib-devhelp==2.0.0
+sphinxcontrib-htmlhelp==2.1.0
+sphinxcontrib-jquery==4.1
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==2.0.0
+sphinxcontrib-serializinghtml==2.0.0
 SQLAlchemy==2.0.32
 threadpoolctl==3.5.0
 tomli==2.0.1
@@ -87,3 +116,4 @@ vine==5.1.0
 wcwidth==0.2.13
 Werkzeug==3.1.3
 wsproto==1.2.0
+WTForms==3.2.1
